requirements.txt
pandas
numpy
luigi
beautifulsoup4
selenium
undetected-chromedriver
google-generativeai